Create BuildResponse interface
