+++ src/main/java/com/heatheryou/httpserver/route/handler/BuildResponse.java
@@ -2,6 +2,6 @@ package com.heatheryou.httpserver.route.handler;
 
 import com.heatheryou.httpserver.Response;
 
-public interface IResponseBuilder {
+public interface BuildResponse {
     Response buildResponse(int statusCode, String[] entityHeaders, String content);
 }
