+++ src/test/java/com/heatheryou/httpserver/ServerTest.java
@@ -1,6 +1,7 @@
 package com.heatheryou.httpserver;
 
 import com.heatheryou.httpserver.route.Router;
+import com.heatheryou.httpserver.route.handler.ResponseBuilder;
 import org.junit.Before;
 import org.junit.Test;
